+++ WatsonWebserver/ReflectionCore.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -17,38 +19,55 @@ namespace WatsonWebserver
         /// <param name="server">Server.</param>
         /// <param name="assembly">Assembly.</param>
         /// <returns>Server.</returns>
-        public static T LoadRoutes<T>(this T server, Assembly assembly = null) where T : Server
-        {
-            var routes = (assembly ?? Assembly.GetCallingAssembly())
+        public static Server LoadRoutes<T>(this T server, Assembly assembly = null) where T : Server
+        { 
+            var staticRoutes = (assembly ?? Assembly.GetCallingAssembly())
                 .GetTypes() // Get all classes from assembly
                 .SelectMany(x => x.GetMethods()) // Get all methods from assembly
-                .Where(IsValidRoute); // Only select methods that are valid routes
+                .Where(IsStaticRoute); // Only select methods that are valid routes
 
-            foreach (var route in routes)
+            var dynamicRoutes = (assembly ?? Assembly.GetCallingAssembly())
+                .GetTypes() // Get all classes from assembly
+                .SelectMany(x => x.GetMethods()) // Get all methods from assembly
+                .Where(IsDynamicRoute); // Only select methods that are valid routes
+             
+            foreach (var staticRoute in staticRoutes)
             {
-                var attribute = route.GetCustomAttributes().OfType<RouteAttribute>().First();
-                server.StaticRoutes.Add(attribute.Method, attribute.Path, route.ToRouteMethod());
+                var attribute = staticRoute.GetCustomAttributes().OfType<StaticRouteAttribute>().First();
+                server.StaticRoutes.Add(attribute.Method, attribute.Path, staticRoute.ToRouteMethod());
+            }
+
+            foreach (var dynamicRoute in dynamicRoutes)
+            {
+                var attribute = dynamicRoute.GetCustomAttributes().OfType<DynamicRouteAttribute>().First();
+                server.DynamicRoutes.Add(attribute.Method, attribute.Path, dynamicRoute.ToRouteMethod());
             }
 
             return server;
         }
          
-        private static bool IsValidRoute(MethodInfo method)
-            => method.GetCustomAttributes().OfType<RouteAttribute>().Any() // Must have the Route attribute
+        private static bool IsStaticRoute(MethodInfo method)
+            => method.GetCustomAttributes().OfType<StaticRouteAttribute>().Any() 
                && method.ReturnType == typeof(Task)
                && method.GetParameters().Length == 1
-               && method.GetParameters().First().ParameterType == typeof(HttpContext); 
+               && method.GetParameters().First().ParameterType == typeof(HttpContext);
+
+        private static bool IsDynamicRoute(MethodInfo method)
+            => method.GetCustomAttributes().OfType<DynamicRouteAttribute>().Any() 
+               && method.ReturnType == typeof(Task)
+               && method.GetParameters().Length == 1
+               && method.GetParameters().First().ParameterType == typeof(HttpContext);
          
         private static Func<HttpContext, Task> ToRouteMethod(this MethodInfo method)
         {
             if (method.IsStatic)
-                return (Func<HttpContext, Task>) Delegate.CreateDelegate(typeof(Func<HttpContext, Task>), method);
+            {
+                return (Func<HttpContext, Task>)Delegate.CreateDelegate(typeof(Func<HttpContext, Task>), method);
+            }
             else
             {
-                object classInstance =
-                    Activator.CreateInstance(method.DeclaringType ?? throw new Exception("Declaring class is null"));
-                return (Func<HttpContext, Task>) Delegate.CreateDelegate(typeof(Func<HttpContext, Task>), classInstance,
-                    method);
+                object instance = Activator.CreateInstance(method.DeclaringType ?? throw new Exception("Declaring class is null"));
+                return (Func<HttpContext, Task>)Delegate.CreateDelegate(typeof(Func<HttpContext, Task>), instance, method);
             }
         }
     }